Move ActivationKeys to component
